unconfirmed_link.vue を react に対応させた
