+++ app/javascript/unconfirmed_link.vue
@@ -1,25 +1,17 @@
 <template lang="pug">
-.card-footer
-  .card-main-actions
-    ul.card-main-actions__items
-      li.card-main-actions__item
-        button.thread-unconfirmed-links-form__action.a-button.is-sm.is-block.is-secondary(
-          @click='openUnconfirmedItems()') {{ label }}
+.react-root
 </template>
 <script>
+import React from 'react'
+import { render } from 'react-dom'
+import UnconfirmedLink from 'components/UnconfirmedLink'
+
 export default {
   props: {
     label: { type: String, required: true }
   },
-  methods: {
-    openUnconfirmedItems() {
-      const links = document.querySelectorAll(
-        '.card-list-item .js-unconfirmed-link'
-      )
-      links.forEach((link) => {
-        window.open(link.href, '_target', 'noopener')
-      })
-    }
+  mounted() {
+    render(<UnconfirmedLink label={this.label} />, this.$el)
   }
 }
 </script>
